Remoção de pastas e scripts que não estavam sendo usados na visão atual. Atualização do código principal
