--- src/vision/object_finder/data/label_map.pbtxt
@@ -1,19 +1,0 @@
-item {
-  id: 1
-  name: 'ball'
-}
-
-item {
-  id: 2
-  name: 'post'
-}
-
-item {
-  id: 3
-  name: 'robot'
-}
-
-item {
-  id: 4
-  name: 'referee'
-}
