--- src/vision/object_finder/pre_trained_models/ssd_edrom/checkpoint
@@ -1,2 +1,0 @@
-model_checkpoint_path: "model.ckpt"
-all_model_checkpoint_paths: "model.ckpt"
