--- src/vision/robocup_cnn_files/backup_robocup/yolov4-tiny-obj.cfg
@@ -1,281 +1,0 @@
-[net]
-# Testing
-#batch=1
-#subdivisions=1
-# Training
-batch=64
-subdivisions=16
-width=416
-height=416
-channels=3
-momentum=0.9
-decay=0.0005
-angle=0
-saturation = 1.5
-exposure = 1.5
-hue=.1
-
-learning_rate=0.00261
-burn_in=1000
-max_batches = 6000
-policy=steps
-steps=4800,5400
-scales=.1,.1
-
-[convolutional]
-batch_normalize=1
-filters=32
-size=3
-stride=2
-pad=1
-activation=leaky
-
-[convolutional]
-batch_normalize=1
-filters=64
-size=3
-stride=2
-pad=1
-activation=leaky
-
-[convolutional]
-batch_normalize=1
-filters=64
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers=-1
-groups=2
-group_id=1
-
-[convolutional]
-batch_normalize=1
-filters=32
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[convolutional]
-batch_normalize=1
-filters=32
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers = -1,-2
-
-[convolutional]
-batch_normalize=1
-filters=64
-size=1
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers = -6,-1
-
-[maxpool]
-size=2
-stride=2
-
-[convolutional]
-batch_normalize=1
-filters=128
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers=-1
-groups=2
-group_id=1
-
-[convolutional]
-batch_normalize=1
-filters=64
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[convolutional]
-batch_normalize=1
-filters=64
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers = -1,-2
-
-[convolutional]
-batch_normalize=1
-filters=128
-size=1
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers = -6,-1
-
-[maxpool]
-size=2
-stride=2
-
-[convolutional]
-batch_normalize=1
-filters=256
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers=-1
-groups=2
-group_id=1
-
-[convolutional]
-batch_normalize=1
-filters=128
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[convolutional]
-batch_normalize=1
-filters=128
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers = -1,-2
-
-[convolutional]
-batch_normalize=1
-filters=256
-size=1
-stride=1
-pad=1
-activation=leaky
-
-[route]
-layers = -6,-1
-
-[maxpool]
-size=2
-stride=2
-
-[convolutional]
-batch_normalize=1
-filters=512
-size=3
-stride=1
-pad=1
-activation=leaky
-
-##################################
-
-[convolutional]
-batch_normalize=1
-filters=256
-size=1
-stride=1
-pad=1
-activation=leaky
-
-[convolutional]
-batch_normalize=1
-filters=512
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[convolutional]
-size=1
-stride=1
-pad=1
-filters=21
-activation=linear
-
-
-
-[yolo]
-mask = 3,4,5
-anchors = 10,14,  23,27,  37,58,  81,82,  135,169,  344,319
-classes=2
-num=6
-jitter=.3
-scale_x_y = 1.05
-cls_normalizer=1.0
-iou_normalizer=0.07
-iou_loss=ciou
-ignore_thresh = .7
-truth_thresh = 1
-random=0
-resize=1.5
-nms_kind=greedynms
-beta_nms=0.6
-
-[route]
-layers = -4
-
-[convolutional]
-batch_normalize=1
-filters=128
-size=1
-stride=1
-pad=1
-activation=leaky
-
-[upsample]
-stride=2
-
-[route]
-layers = -1, 23
-
-[convolutional]
-batch_normalize=1
-filters=256
-size=3
-stride=1
-pad=1
-activation=leaky
-
-[convolutional]
-size=1
-stride=1
-pad=1
-filters=21
-activation=linear
-
-[yolo]
-mask = 0,1,2
-anchors = 10,14,  23,27,  37,58,  81,82,  135,169,  344,319
-classes=2
-num=6
-jitter=.3
-scale_x_y = 1.05
-cls_normalizer=1.0
-iou_normalizer=0.07
-iou_loss=ciou
-ignore_thresh = .7
-truth_thresh = 1
-random=0
-resize=1.5
-nms_kind=greedynms
-beta_nms=0.6
