--- src/vision/robocup_videos/valid.txt
@@ -1,136 +1,0 @@
-data/obj/val_98.jpg
-data/obj/val_120.jpg
-data/obj/val_66.jpg
-data/obj/val_41.jpg
-data/obj/val_118.jpg
-data/obj/val_36.jpg
-data/obj/val_57.jpg
-data/obj/val_124.jpg
-data/obj/val_88.jpg
-data/obj/val_58.jpg
-data/obj/val_111.jpg
-data/obj/val_101.jpg
-data/obj/val_17.jpg
-data/obj/val_16.jpg
-data/obj/val_20.jpg
-data/obj/val_6.jpg
-data/obj/val_52.jpg
-data/obj/val_45.jpg
-data/obj/val_73.jpg
-data/obj/val_23.jpg
-data/obj/val_18.jpg
-data/obj/val_135.jpg
-data/obj/val_27.jpg
-data/obj/val_50.jpg
-data/obj/val_34.jpg
-data/obj/val_51.jpg
-data/obj/val_123.jpg
-data/obj/val_55.jpg
-data/obj/val_97.jpg
-data/obj/val_104.jpg
-data/obj/val_99.jpg
-data/obj/val_61.jpg
-data/obj/val_77.jpg
-data/obj/val_103.jpg
-data/obj/val_53.jpg
-data/obj/val_125.jpg
-data/obj/val_39.jpg
-data/obj/val_56.jpg
-data/obj/val_102.jpg
-data/obj/val_63.jpg
-data/obj/val_24.jpg
-data/obj/val_76.jpg
-data/obj/val_91.jpg
-data/obj/val_1.jpg
-data/obj/val_30.jpg
-data/obj/val_113.jpg
-data/obj/val_15.jpg
-data/obj/val_131.jpg
-data/obj/val_7.jpg
-data/obj/val_132.jpg
-data/obj/val_11.jpg
-data/obj/val_2.jpg
-data/obj/val_80.jpg
-data/obj/val_115.jpg
-data/obj/val_79.jpg
-data/obj/val_134.jpg
-data/obj/val_64.jpg
-data/obj/val_31.jpg
-data/obj/val_28.jpg
-data/obj/val_83.jpg
-data/obj/val_82.jpg
-data/obj/val_95.jpg
-data/obj/val_47.jpg
-data/obj/val_136.jpg
-data/obj/val_32.jpg
-data/obj/val_87.jpg
-data/obj/val_112.jpg
-data/obj/val_13.jpg
-data/obj/val_126.jpg
-data/obj/val_78.jpg
-data/obj/val_128.jpg
-data/obj/val_12.jpg
-data/obj/val_109.jpg
-data/obj/val_49.jpg
-data/obj/val_117.jpg
-data/obj/val_10.jpg
-data/obj/val_69.jpg
-data/obj/val_44.jpg
-data/obj/val_93.jpg
-data/obj/val_114.jpg
-data/obj/val_94.jpg
-data/obj/val_107.jpg
-data/obj/val_38.jpg
-data/obj/val_72.jpg
-data/obj/val_133.jpg
-data/obj/val_26.jpg
-data/obj/val_19.jpg
-data/obj/val_67.jpg
-data/obj/val_90.jpg
-data/obj/val_65.jpg
-data/obj/val_62.jpg
-data/obj/val_110.jpg
-data/obj/val_81.jpg
-data/obj/val_100.jpg
-data/obj/val_43.jpg
-data/obj/val_4.jpg
-data/obj/val_105.jpg
-data/obj/val_84.jpg
-data/obj/val_60.jpg
-data/obj/val_14.jpg
-data/obj/val_59.jpg
-data/obj/val_35.jpg
-data/obj/val_68.jpg
-data/obj/val_5.jpg
-data/obj/val_74.jpg
-data/obj/val_33.jpg
-data/obj/val_71.jpg
-data/obj/val_89.jpg
-data/obj/val_121.jpg
-data/obj/val_129.jpg
-data/obj/val_9.jpg
-data/obj/val_96.jpg
-data/obj/val_116.jpg
-data/obj/val_106.jpg
-data/obj/val_92.jpg
-data/obj/val_37.jpg
-data/obj/val_42.jpg
-data/obj/val_48.jpg
-data/obj/val_86.jpg
-data/obj/val_40.jpg
-data/obj/val_108.jpg
-data/obj/val_119.jpg
-data/obj/val_85.jpg
-data/obj/val_3.jpg
-data/obj/val_122.jpg
-data/obj/val_21.jpg
-data/obj/val_70.jpg
-data/obj/val_25.jpg
-data/obj/val_29.jpg
-data/obj/val_54.jpg
-data/obj/val_46.jpg
-data/obj/val_75.jpg
-data/obj/val_8.jpg
-data/obj/val_22.jpg
-data/obj/val_130.jpg
-data/obj/val_127.jpg
